VerifyMacros

Documentation for VerifyMacros.jl.

Overview

This package provides convenient macros that verify whether given variables have given properties. These macros apply to types, fields, keys, properties, axes, and supertypes.

For instance,

    @verifytype x Float64

verifies that x is a Float64 and prints a convenient error message, otherwise. An example:

julia> using VerifyMacros
julia> x = 3.03.0
julia> @verifytype x IntERROR: TypeError: x is of type Float64; was expecting a Int64
#= REPL[3]:1 =#

If desirable, one can change the description of the variable by providing a third argument, thus:

julia> using VerifyMacros
julia> x = 3.03.0
julia> @verifytype x Int "Some variable"ERROR: TypeError: Some variable is of type Float64; was expecting a Int64
#= REPL[3]:1 =#

There are multi-argument versions of the same. For instance, one can do

julia> using VerifyMacros
julia> x = 3.03.0
julia> y = "blooper""blooper"
julia> @verifytypes (x,AbstractFloat) (y,Symbol,why)ERROR: TypeError: why is of type String; was expecting a Symbol
#= /home/runner/work/VerifyMacros.jl/VerifyMacros.jl/src/verify.jl:138 =#

List of macros

VerifyMacros.@verifyaxes

VerifyMacros.@verifyaxes_list

VerifyMacros.@verifyfield

VerifyMacros.@verifyfields

VerifyMacros.@verifykey

VerifyMacros.@verifykeys

VerifyMacros.@verifyproperties

VerifyMacros.@verifyproperty

VerifyMacros.@verifysupertype

VerifyMacros.@verifysupertypes

VerifyMacros.@verifytype

VerifyMacros.@verifytypes

VerifyMacros.@verifyaxes — Macro

@verifyaxes(array, expected_axes[, name])

Verify that array has the specified expected_axes. Throws a descriptive error if axes don't match. Optionally provide a custom name for the array in error messages.

VerifyMacros.@verifyaxes_list — Macro

@verifyaxes_list((array1, axes1), (array2, axes2), ...)

Batch verify multiple axes checks. Each argument must be a tuple of (array, expected_axes[, name]). Equivalent to multiple @verifyaxes calls.

VerifyMacros.@verifyfield — Macro

@verifyfield(Type, field[, name])

Verify that Type has the specified field. Throws a descriptive error if the field is missing. Optionally provide a custom name for the type in error messages.

VerifyMacros.@verifyfields — Macro

@verifyfields((Type1, field1), (Type2, field2), ...)

Batch verify multiple field checks. Each argument must be a tuple of (Type, field[, name]). Equivalent to multiple @verifyfield calls.

VerifyMacros.@verifykey — Macro

@verifykey(dict, key[, name])

Verify that dict contains the specified key. Throws a descriptive error if the key is missing. Optionally provide a custom name for the dictionary in error messages.

VerifyMacros.@verifykeys — Macro

@verifykeys((dict1, key1), (dict2, key2), ...)

Batch verify multiple key checks. Each argument must be a tuple of (dict, key[, name]). Equivalent to multiple @verifykey calls.

VerifyMacros.@verifyproperties — Macro

@verifyproperties((object1, property1), (object2, property2), ...)

Batch verify multiple property checks. Each argument must be a tuple of (object, property[, name]). Equivalent to multiple @verifyproperty calls.

VerifyMacros.@verifyproperty — Macro

@verifyproperty(object, property[, name])

Verify that object has the specified property. Throws a descriptive error if the property is missing. Optionally provide a custom name for the object in error messages.

VerifyMacros.@verifysupertype — Macro

@verifysupertype(Type, SuperType[, name])

Verify that Type is a subtype of SuperType. Throws a descriptive error if the subtype relation fails. Optionally provide a custom name for the type in error messages.

VerifyMacros.@verifysupertypes — Macro

@verifysupertypes((Type1, SuperType1), (Type2, SuperType2), ...)

Batch verify multiple subtype relations. Each argument must be a tuple of (Type, SuperType[, name]). Equivalent to multiple @verifysupertype calls.

VerifyMacros.@verifytype — Macro

@verifytype(value, Type[, name])

Verify that value is of type Type. Throws a descriptive error if the type check fails. Optionally provide a custom name for the value in error messages.

VerifyMacros.@verifytypes — Macro

@verifytypes((value1, Type1), (value2, Type2), ...)

Batch verify multiple type checks. Each argument must be a tuple of (value, Type[, name]). Equivalent to multiple @verifytype calls.