VerifyMacros

Documentation for VerifyMacros.jl.

Overview

This package provides convenient macros that verify whether given variables have given properties. These macros apply to types, fields, keys, properties, axes, supertypes, equality, length, size, file/directory existence, and generic conditions.

For instance,

    @verifytype x Float64

verifies that x is a Float64 and prints a convenient error message, otherwise. An example:

julia> using VerifyMacros
julia> x = 3.03.0
julia> @verifytype x IntERROR: TypeError: x is of type Float64; was expecting a Int64
#= REPL[3]:1 =#

If desirable, one can change the description of the variable by providing a third argument, thus:

julia> using VerifyMacros
julia> x = 3.03.0
julia> @verifytype x Int "Some variable"ERROR: TypeError: Some variable is of type Float64; was expecting a Int64
#= REPL[3]:1 =#

There are multi-argument versions of the same. For instance, one can do

julia> using VerifyMacros
julia> x = 3.03.0
julia> y = "blooper""blooper"
julia> @verifytypes (x,AbstractFloat) (y,Symbol,why)ERROR: TypeError: why is of type String; was expecting a Symbol
#= REPL[4]:1 =#

List of macros

VerifyMacros.@verifyaxes

VerifyMacros.@verifyaxesm

VerifyMacros.@verifyequal

VerifyMacros.@verifyequals

VerifyMacros.@verifyfield

VerifyMacros.@verifyfields

VerifyMacros.@verifyin

VerifyMacros.@verifyins

VerifyMacros.@verifyisdir

VerifyMacros.@verifyisdirs

VerifyMacros.@verifyisfile

VerifyMacros.@verifyisfiles

VerifyMacros.@verifykey

VerifyMacros.@verifykeys

VerifyMacros.@verifylength

VerifyMacros.@verifylengths

VerifyMacros.@verifyproperties

VerifyMacros.@verifyproperty

VerifyMacros.@verifysize

VerifyMacros.@verifysizes

VerifyMacros.@verifysupertype

VerifyMacros.@verifysupertypes

VerifyMacros.@verifytrue

VerifyMacros.@verifytrues

VerifyMacros.@verifytype

VerifyMacros.@verifytypes

VerifyMacros.@verifyaxes — Macro

@verifyaxes(array, expected_axes[, name])

Verify that array has the specified expected_axes. Throws a descriptive error if axes don't match. Optionally provide a custom name for the array in error messages.

VerifyMacros.@verifyaxesm — Macro

@verifyaxesm((array1, axes1), (array2, axes2), ...)

Batch verify multiple axes checks. Each argument must be a tuple of (array, expected_axes[, name]). Equivalent to multiple @verifyaxes calls.

VerifyMacros.@verifyequal — Macro

@verifyequal(value, expected[, name])

Verify that value is equal to expected. Throws a descriptive error if not equal. Optionally provide a custom name for the value in error messages.

VerifyMacros.@verifyequals — Macro

@verifyequals((value1, expected1), (value2, expected2), ...)

Batch verify multiple equality checks. Each argument must be a tuple of (value, expected[, name]). Equivalent to multiple @verifyequal calls.

VerifyMacros.@verifyfield — Macro

@verifyfield(Type, field[, name])

Verify that Type has the specified field. Throws a descriptive error if the field is missing. Optionally provide a custom name for the type in error messages.

VerifyMacros.@verifyfields — Macro

@verifyfields((Type1, field1), (Type2, field2), ...)

Batch verify multiple field checks. Each argument must be a tuple of (Type, field[, name]). Equivalent to multiple @verifyfield calls.

VerifyMacros.@verifyin — Macro

@verifyin(element, collection[, name])

Verify that element is in collection. Throws a descriptive error if the check fails. Optionally provide a custom name for the element in error messages.

VerifyMacros.@verifyins — Macro

@verifyins((element1, Collection1), (element2, Collection2), ...)

Batch verify multiple membership checks. Each argument must be a tuple of (element, Collection[, name]). Equivalent to multiple @verifyin calls.

VerifyMacros.@verifyisdir — Macro

@verifyisdir(path[, name])

Verify that path is an existing directory. Throws a descriptive error if not. Optionally provide a custom name for the directory in error messages.

VerifyMacros.@verifyisdirs — Macro

@verifyisdirs((path1), (path2), ...)

Batch verify multiple directory existence checks. Each argument must be a tuple of (path[, name]). Equivalent to multiple @verifyisdir calls.

VerifyMacros.@verifyisfile — Macro

@verifyisfile(path[, name])

Verify that path is an existing file. Throws a descriptive error if not. Optionally provide a custom name for the file in error messages.

VerifyMacros.@verifyisfiles — Macro

@verifyisfiles((path1), (path2), ...)

Batch verify multiple file existence checks. Each argument must be a tuple of (path[, name]). Equivalent to multiple @verifyisfile calls.

VerifyMacros.@verifykey — Macro

@verifykey(dict, key[, name])

Verify that dict contains the specified key. Throws a descriptive error if the key is missing. Optionally provide a custom name for the dictionary in error messages.

VerifyMacros.@verifykeys — Macro

@verifykeys((dict1, key1), (dict2, key2), ...)

Batch verify multiple key checks. Each argument must be a tuple of (dict, key[, name]). Equivalent to multiple @verifykey calls.

VerifyMacros.@verifylength — Macro

@verifylength(collection, len[, name])

Verify that collection has length len. Throws a descriptive error if the length check fails. Optionally provide a custom name for the collection in error messages.

VerifyMacros.@verifylengths — Macro

@verifylengths((col1, len1), (col2, len2), ...)

Batch verify multiple length checks. Each argument must be a tuple of (collection, length[, name]). Equivalent to multiple @verifylength calls.

VerifyMacros.@verifyproperties — Macro

@verifyproperties((object1, property1), (object2, property2), ...)

Batch verify multiple property checks. Each argument must be a tuple of (object, property[, name]). Equivalent to multiple @verifyproperty calls.

VerifyMacros.@verifyproperty — Macro

@verifyproperty(object, property[, name])

Verify that object has the specified property. Throws a descriptive error if the property is missing. Optionally provide a custom name for the object in error messages.

VerifyMacros.@verifysize — Macro

@verifysize(array, size[, name])

Verify that array has size size. Throws a descriptive error if the size check fails. Optionally provide a custom name for the array in error messages.

VerifyMacros.@verifysizes — Macro

@verifysizes((arr1, size1), (arr2, size2), ...)

Batch verify multiple size checks. Each argument must be a tuple of (array, size[, name]). Equivalent to multiple @verifysize calls.

VerifyMacros.@verifysupertype — Macro

@verifysupertype(Type, SuperType[, name])

Verify that Type is a subtype of SuperType. Throws a descriptive error if the subtype relation fails. Optionally provide a custom name for the type in error messages.

VerifyMacros.@verifysupertypes — Macro

@verifysupertypes((Type1, SuperType1), (Type2, SuperType2), ...)

Batch verify multiple subtype relations. Each argument must be a tuple of (Type, SuperType[, name]). Equivalent to multiple @verifysupertype calls.

VerifyMacros.@verifytrue — Macro

@verifytrue(condition[, name])

Verify that condition is true. Throws a descriptive error if it evaluates to false. Optionally provide a custom name for the condition in error messages.

VerifyMacros.@verifytrues — Macro

@verifytrues((condition1), (condition2), ...)

Batch verify multiple conditions. Each argument must be a tuple of (condition[, name]). Equivalent to multiple @verifytrue calls.

VerifyMacros.@verifytype — Macro

@verifytype(value, Type[, name])

Verify that value is of type Type. Throws a descriptive error if the type check fails. Optionally provide a custom name for the value in error messages.

VerifyMacros.@verifytypes — Macro

@verifytypes((value1, Type1), (value2, Type2), ...)

Batch verify multiple type checks. Each argument must be a tuple of (value, Type[, name]). Equivalent to multiple @verifytype calls.